Trying to refactor interfaces so they make more sense.
A worker is passed an application. It's responsible for setting up
the connection to the scheduler and providing that interface to
an application through a Client object. The application invokes
the client object to submit jobs. The worker is responsible for
reading jobs from the server and invoking them on the application.
Exactly how this interface will work out is an interesting question.
It will necessitate a mapping between string identifiers and functions
as well as data.

The concurrency of the client object is an interesting and open question.
I.e. should it spawn threads or should it directly invoke. It seems to
me that the client should *not* have threads. The worker is responsible
for reading jobs, while the application code (invoked from the worker
through jobs) is responsible for writing jobs.

I think loadApp makes sense as the function called on one and only
one worker, to set up the simulation.