We don't want to pass ports directly to the worker; that requires the worker to have the logic and control of establishing its communication channel to the scheduler. A better approach is to pass in the communication client. That way, if we want, we can instantiate a subclass which happens to read from a file, or some other I/O mechanism. Otherwise, we have to encode every such variation inside the Worker.
The order of setup should be this:

create SchedulerClient
start client to make sure comms are up
create app
load app so it can read in all of its info and state and such
create worker with app
tell worker to run with client

The worker then invokes start() on the app, so it can send
initial commands (taking loaded state and turning it into commands,
data elements, etc. for scheduler).

Finally, the worker enters its core loop of pulling off commands
and executing them.