big commit, fixed a couple of issues: 1. The sendCommand and receiveCommand used to be blocking. read_until and getline are blocking and had to change it so if the command is received partially or there are no comming commands the process is not blocked.
2. The scheduler/worker used to crash when one side was down.
It was because of the broken pipe after one side was dead.
It is now working properly, the dead worker should be cleaned
after it is not responsive for a while (will do it later.).

3. Now we can launch a scheduler and a bunch of worker,
either from /worker or /test/worker_commands. They connect
scheduler and you can bring any of them down without affecting
the rest of the system.