Refactored scheduler server interface: .cc still needs updates.
Added a worker_id_t type and all commands now have an associated worker id.
This is a piece of metadata. In the case of received commands, it's the source
worker, in the case of sent commands, it's the destination worker. Since the
Scheduler Server clusters many commands into a single read call, it's important
(at least from a logging/debugging standpoint) to know which workers they came
from. Open question whether the worker ID is actually known to the worker.