found and fixed the bug, we officially have water application over two separate workres :)
it was the DeSerialize method of the scalar_data that did not allocate memory
for the new data, so remote receive job would fail at destroying stage. It used
to work because we tried partitioed multiple jobs on one worker and never had
remote copy (all local copy). scheduler alternate is great for intense
evaluation of the code.