fixed the bug in job_manger: job is not ready to assign before parent is done (in addition to before set), also set the kernel job to done so that main job could be spawned.
this causes problem specially for the first job that parent spawnes, since it
has no before set and parent is implicitly in its before set. the child needs
to wait for the parent to be done to avoid destroying necessary versions for
the unknown siblings.  this was causing the lost version ERROR in water
multiple when I tried to split the first job in loop iteration.