Omid spent a while today complaining about a bug that he found with scoped locks in nimbus.  Essentially, a common error with scoped locks is to construct the scoped lock, but fail to assign it to a variable. Eg.,
   Mutex m;
   scoped_lock(m);

instead of

   Mutex m;
   scoped_lock sl (m);

In the first case, the block of code instantly constructs and destroys
a scoped lock object.  Subsequent code that is executed is unguarded
by the mutex.  In contrast, in the second example, code executed after
constructing the scoped lock is guarded until sl goes out of scope.

This script tries to check for this error.  It accepts a series of
folders and examines all c++ files in that folder and its subfolders
looking for the error described above.  See docstring at top of file
for more info, including false positives.