found and fixed the bug why sometimes after rewinding workers were blocked.
fix: needed to clean the list job read and last job write for all instances in
addition to version reset.

problem: it was causing obsolete jobs appear in before sets. controller did not
clean scheduler created jobs (copy/create). Although worker ignores them if it
cannot find them in its job graph (so it worked some times), but there is a
corner case for remote copy receive jobs when worker graph is cleared and then
data is received. In this case worker adds the node again to the graph and so
the obsolete job does not get cleared for the new coming jobs.

worker memory leak: currently I fixed the controller to remove all obsolete
jobs from data manager and so they do not appear in before sets. but worker
could still recreate pending nodes in its job graph and that is a memory leak.
will fix this worker's problem but it does not affect correctness.