added a hack to fix the problem with projection_loop_iteration template:
problem: the first pli has a diffrent relative position compared to the rest of
them. the first one comes after projection main and the rest of them come after
another pli. since we memoize for a middle one, the binding template does not
work for the first one.

now 8 workers and 64 partitions works on my machine localy, but scheduler has
to have single thread for assignment if binding template is active.