fixed a bug in data exchanger when header buffer goes out of scope.
with asynch write the buffer should remain valid until the handler gets called,
so allocating header buffer on stack as local variable could cause
undeterministic behaviour. I write the header as a field member of serialized
data and then get a reference to it to pass to the async write.