generalized how the iteration boundary is marked for stats collections.
if name of the job or template includes __MARK_STAT, then controller triggers
the stat collection procedure. updated the job names for water_multiple,
stencil_1d, and logistic regression accordingly. scripts are updated
accordingly, and made sure that they are backward compatible to be able to
still process the old collected logs.