fixed the rare race in execution template:
in the rare case that a new template instantiation command is received from
controller before the current template finishes execution, it only saves the
meta data and instantiates it when the last job of the previous template is
marked done. this way it is guaranteed that two templates never use the same
data structure without actually duplicating the data structure.