changed the order of reduction levels in logistic regresssion:
for multi-tenant experiment when switching from 100 to 50 workers, the
controller ended up adding so many copy jobs for updating per partition weights
from the evicted workers at the beginning of the iteration. Also it added extra
copy jobs at the end for cascaded binding to refresh the values it brought in
from evicted workers (because they get updated in the block). I solved this
problem by changing the order of operations in the block

from: gradient -> reduce_l1 -> reduce_l2 -> synch
to:   synch -> gradient -> reduce_l1 -> reduce_l2

This way since the synching worker is not evicted then there is no need to add
extra copy jobs to bring in the data (even if the synching worker is evicted,
synch reads in fewer data ids compared to reduce_l1, so fewer copy jobs). Also
it ends up adding no copy for cascaded binding, because the scartch weights
gets updated naturally in the loop (or again if the synching worker is evicted,
fewer copies)