MAJOR COMMIT: added a version of boost and protobuf libraries locally as external dependencies of nimbus:
1. upopn make (by default), a version of boost and protobuf is installed
locally for nimbus. nimbus is independent of system wide installed libraries,
all it needs will be installed locally.

2. since we do not need all boost binaries only a subset of it is installed for
nimbus use.  all code is compiled and linekd with the locally installed protoc
and google/protobud headers, this removes the possible inconsistency in
protobuf installations.

3. added Makeinclude.cmake for physbam applications to include in their cmake
file so that they compile with the same libraries and header files of nimbus
and protobuf as nimbus.

next, I will add a bootstrap to allow the user whether to install libraries
locally or use already installed system-wide libraries.